Update patch set 3
Patch Set 3:

There's a difference between DropdownWidget (which is our perfect design and API unencumbered by legacy stuff) and DropdownInputWidget (which is what we can do in practice when we have to be compatible with HTTP form submission and when we have to be able to generate the widget using only HTML+CSS).

(It is unfortunate that the naming of InputWidget subclasses is confusing, I too wish we'd had the foresight to prevent that three years ago.)

You can use either DropdownInputWidget or DropdownWidget in your JavaScript code. But in PHP code, only DropdownInputWidget is available, for the reasons I mentioned. HTMLForm (which by the way is another layer of abstraction and limitations) therefore uses DropdownInputWidget.

If you look at DropdownInputWidget demos (either in the JS version you linked, or in PHP: https://doc.wikimedia.org/oojs-ui/master/demos/demos.php), none of them have a custom label/placeholder.

I'm not sure what you mean by "close the gap between PHP and JavaScript". PHP OOUI provides a subset of the API provided by JS OOUI. It does not provide functionality that is impossible to implement.

Patch-set: 3