Redirect creation: Don't set EDIT_NEW on edit if the page exists
The purpose of this check was to allow creating redirects
as new pages, but only if a page existed before. But we
also set EDIT_NEW in case an existing page has deleted
revisions.
Fix this by just setting EDIT_NEW if the page doesn't exist
but it has existed before.

Bug: T163748
Change-Id: I1e2edd2089044e54c3e96b9306f54408615863c3