Update patch set 1
Patch Set 1:

Checking isDeleted was actually intentional: we don't want redirects to be created from scratch. There should be no reason to do this, ever. Redirects are the result of merging two items. If an item never existed, there is no reason to redirect the respective ID.

It's unclear to me how the error is triggered when deleted revisions are present. That happens if the page exists and has deleted revisions, not if the page was deleted, right?

It seems this patch allows more than is necessary to fix the issue at hand. In any case, we should prevent redirects being created "from scratch".

Patch-set: 1
Label: Code-Review=0