Update patch set 2
Patch Set 2:

> How urgent is this?

I don't think it is super urgent, given this code is in place since 2015 and we only know about this problem as of this week. Would be nice to get it merged for the branch, though.

Patch-set: 2
Reviewer: Hoo man <50@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0