Update patch set 1
Patch Set 1: Code-Review+2

Tested manually. With "return false" the form just says "There are problems with some of your input" without specifying the field in which the error happened. With the error message the field is highlighted. The only minor issue I have with this is that the error message is not perfect. It says "Invalid data type specified" when there is *no* data type specified. But this really is a nitpick.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2