Move setServices in specials to constructor params
Various special pages had set*Services() methods, which made it possible
for tests to override certain services the special pages used. However,
the better approach to allow this is to add these services to the
constructor parameters. This commit moves the services to the
constructor parameters and removes the set*Services() methods.

Bug: T162540
Change-Id: Ieeeb3740d32b9c5de2f00a56197d5ff6efc63aa3