Update patch set 1
Patch Set 1:

(1 comment)

Thanks for the patch! Yes, this is indeed what was intended by the ticket.

The next step would be to stop using the default services in the tests, and use mocks instead. That should allow us to improve the tests.

There is one minor problem, mentioned inline.

Patch-set: 1
Label: Verified=0