Update patch set 2
Patch Set 2:

This change depends entirely on weird code in core. How would you test this without creating an actual entity with a deleted revision? Sure, this can be done, but is this really worth it? On the other hand, if you mock all services you can make the lookup return a dummy Title object. All the test can do then is to make sure the Title::isDeletedQuick() method is not called without calling Title::exists() first. Such a test would encode weird behavior of core, and not be of much value, in my humble opinion.

I think we should merge this as it is and deprecate the bad Title::isDeleted and isDeletedQuick methods.

Patch-set: 2