Update patch set 2
Patch Set 2:

I’m not sure how I could do that – if I add constructor parameters to, say, SpecialNewItem, I also have to add them to SpecialNewEntity because otherwise SpecialNewItem doesn’t know what to do with them; but then I also have to add them to SpecialNewProperty, since that still has to call its superclass constructor (SpecialNewEntity) correctly.

Perhaps I could add the parameters to SpecialNewItem and then call setServices() in its constructor, and then add the parameters to SpecialNewEntity in the next change, but that feels weird: I would be adding the setServices() call only to remove it in the next change. Is that really better?

Patch-set: 2