Update patch set 7
Patch Set 7: Verified-1 -Code-Review

(1 comment)

I was not aware I was blocking this for so long. It looks like my main concern got resolved, and all this patch does now is moving existing code around, without changing much. The effect of this patch is that the discussed code is now in lib. This means lib now depends on repo and client, which still feels like we did not gained anything. We replaced one bad dependency with an other one. Will we ever be able to solve this, I wonder?

Needs a rebase now.

Patch-set: 7
Label: Verified=-1