Update patch set 7
Patch Set 7:

> The effect of this patch is that the discussed code is now in lib. This means lib now depends on repo and client, which still feels like we did not gained anything. 

In terms of inter-module dependency, this indeed does not help. It does remove a cross-dependency between classes though, which is much more important. To word it differently: the inter-module dependency is just about the directory the file is in. There is no dependency in code.

As to resolving the module cross-dependency: we discussed that in the context of <https://docs.google.com/document/d/16kPV3Y3YNF-oLV3ZY7hm67c8M3Xm7xoYxGx-b11HMfM/edit>. I imagine part of the bigger solution will be a new centralized services module.

Patch-set: 7