Update patch set 7
Patch Set 7:

I would merge this patch, but it needs a manual rebase first (I will try to do this the next days, if it helps), and I'm not sure about the hook. It looks like some places in the code that called the hook before don't do this any more, which concerns me a bit.

Patch-set: 7