Use Wikibase's CodeSniffer instead of MediaWiki's
This is a test for the new https://github.com/wmde/WikibaseCodeSniffer.
Note that not much is left in the local phpcs.xml rule set, except very
few actual exceptions from the base rule set.

The checks executed are the same before and after this patch, with one
intentional exception: I removed Squiz.Classes.DuplicateProperty, which
turns out to be a JS-only sniff. I think I added it once by mistake.

Change-Id: Ifed1a4ecd49213f1c0a7fda1fc721b84c5799e0a