Update patch set 9
Patch Set 9:

> a) Will this still register all custom entity types in all cases that did this before?

The hook is still run, but it's now run earlier. If WikibaseRepo or WikibaseClient get initialized before the relevant extension hada chance to register the hook handler, the entity types will not be registered. I don't see how that could happen though.

> b) It looks like the hook is now run in many more situations, which might be a performance issue.

The hooks are triggered max 3 times: in WikibaseRepo::newInstance,  WikibaseClient::newInstance, and in WikibaseClient.default.php by the callback for $defaults['entityNamespaces'] (but only if the default isn't overwritten).

Previously, it was run max 2 times (one for the client factory, and once for the repo factory).

Caching can be added, but I don't see a nice place for it, and I don't think it's worth it: The hook is run one more time, it typically contains only trivial code (no I/O), and if we really want to avoid it, we can just set entityNamespaces explicitly in the client settings.

Patch-set: 9