Rename getExternalFormatEntityDeserializer to getAllTypesEntityDeserializer
These methods are a constant source of confusion. That "external" vs.
"internal" difference is almost irrelevant (the only difference is that
an "internal format" deserializer must support all old serialization
formats, which is optional – but not forbidden – for an "external format"
deserializer). What's much more important is to know what entity types an
"entity (de)serializer" supports.#

Relevant to know: Only Item and Property types are supported by the old
"internal format" deserializer. New entity types are never added to this
legacy deserializer, and should not.

I split this from Ic126c46.

Bug: T160439
Change-Id: I7704143f808afcfae59d93c1b45b11fad0a15382