Avoid duplicate and optional code in CreateRedirectTest
This is a direct follow up done while reviewing the changes done in
I1e2edd2. The main goal of this patch is to avoid duplicating the array
with the "BC", "Types", and "Strip" keys, as well as some other duplicate
code. I'm also trying to make method argument *not* optional because
doing so tends to make code less confusing.

Bug: T163748
Change-Id: I4615003d658300ba58d123750036966013249e5b