Rename …StorageFrontendUrl settings to …StorageBaseUrl
This rename was a PM decision when we started exposing these settings
via the API.

This patch obviously changes a public interface, but does not have an
effect, as far as I can see. These two settings are still brand new and
not used anywhere. The defaults are …://commons.wikimedia.org/…. We
will most probably never change these settings on our clusters.

Change-Id: I1bc4f578fe7e219f3e9e0c2ebedbd3d65784a7e9