Use a consistent code style to check for flags in a bit field
Our current code base uses a bunch of different code styles to check
for flags in a bit field:

* ( $var & CONST ) === CONST
* ( $var & CONST ) !== 0
* ( $var & CONST ) > 0
* (bool)( $var & CONST )
* $var & CONST

I find, for example, the first one quite confusing. (Can you quickly
say if it checks if a flag is set or not set?) I also find it confusing
to have so many styles next to each other.

This patch proposes:

* Stick to the "!== 0" style in case it's critical to have a boolean
  result. This is more specific than a vague "> 0". It avoids repeating
  the const name, which is a source of mistakes. It's also close to other
  styles we already use, for example "!== null" or "=== []" for empty
  array checks.
* In an if() use the shortest possible "if ( $var & CONST )". This is
  what most code already does.

This patch directly follows the comments I wrote in I5fc2557.

Change-Id: I7f7bd6a445f6b0c3b9b33ee7d0d7e6adb03209c0