Update patch set 13
Patch Set 13: Code-Review+1

Yeah, I believe that if we say we drop "start-row" option from the maint script, there is even no need to have this as a field in PropertyInfoTableBuilder. Just have method's local variable, initially always 1.

Actually, as now this modifies the value of the class property, the behaviour of the class becomes a bit odd, if I get it correct. So when you call ProprtyInfoTableBuilder::rebuildPropertyInfo it will update all the props. This is fine. Then when running the same method on the same instance again, it will just skip all records (possibly not skipping some new rows added in the meanwhile), right?
Having this just as a method variable would simply make it process all records on each call, which would probably be simplest thing to do.

Patch-set: 13
Label: Code-Review=+1