DNM: DEMONSTRATION ONLY!!!
Allow extensions to register their own services.

See an example of this in action: TODO

This also moves wiring file location from client settings
to its own file. Among other things, this make it easier
to deal with in hook, as there is not need to pass the
whole client settings array to hook handlers.

This is a quick'n'dirty solution made in order to demonstrate
a possible approach that could lead to resolving the Repo/Client
cross dependency

This patch is just a sloppy and clunky draft. Please do not be
distracted by nasty details, and focus on the general idea.

Note: this is certainly not the only possible way to allow
extension to have their own services that access multiple
repositories etc.
Possible alternative might be e.g. using ServiceWiring registration
provided by MediaWiki core?

Change-Id: I98081a5df4e143c9b268bd5d21fb09ba433e861f