Update patch set 3
Patch Set 3: Code-Review+1

I checked and the old has… setting is already live, so having the fallback is actually critical for the moment. Can be ditched later, when the switch to the new column is done and the entire setting is ready to die.

Patch-set: 3
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1