Update patch set 3
Patch Set 3: Code-Review+1

(3 comments)

Looks like the rename would work fine, especially given the fallback to the old setting name.

I just wonder if it is actually better to rely on evaluation-to-bool magic, so I'd like to hear from Thiemo as he seems to prefer this over the strict bool.
I don't it is much of an issue as this setting is supposed to go in the near future, but in principle I wonder if this is the best practice.

Patch-set: 3
Label: Code-Review=+1