Move EntityDataRetrievalServiceFactory and related classes to Data Access component
The new component is (for now) a separate directory in Wikibase.git.

EntityDataRetrievalServiceFactory interface, DispatchingServiceFactory,
and RepositoryServiceContainer, plus their related wiring files are
moved to component's directory.

Possible renaming of existing classes/interfaces etc would
happen separately to avoid mistakes.

In principle no code in DataAccess component should rely on
services, top-level factories etc from Repo and/or Client components.
There is a PHPUnit test in place to assure this restriction
is not violated. Currently there is a single dependency on
WikibaseClient in RepositoryServiceContainer. It is temporarily
allowed until all relevant refactoring is not finished.

Bug: T164737
Change-Id: I6acdf4420fe6be8cc281a4f6a49e6dc6f6cd58f1