Use __DIR__ in all resource loader hooks
This "resources….php" file was the only one where the __DIR__ constant
was missing.

The code style fix to the other file is unrelated, but does not need
it's own patch in my opinion. Feel free to disagree. Touching both files
was motivated by experimenting with the new MediaWiki CodeSniffer 0.8
release.

Change-Id: I940e89c8ea4cfeb428eb210f3ee9a8c5df5cfca2