Update patch set 2
Patch Set 2: Code-Review+1

The error in PS1 was expected. The globe was always validated via ValidatorBuilders::buildCoordinateValidators, and empty strings where always rejected. There was a test for this that can't work any more, because now it is not even possible to create a GlobeCoordinateValue with an empty string.

I'm not sure what the -2 was for (merging this was already blocked by the "verified -1"). I think this is fine to be merged now.

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1