Update patch set 2
Patch Set 2: Code-Review+1

(5 comments)

Code is almost trivial (which is a good sign, seems the existing code we build here is really good). We could merge this as it is, but I would like us to improve some documentation first. Can you do this please?

No -1 because this edit request is not critical at all, and others might disagree.

Patch-set: 2
Label: Code-Review=+1