Use WikibaseServices in Client and Repo instead of EntityDataRetrievalServiceFactory
WikibaseServices currently only provide services that
EntityDataRetrievalServiceFactory also used to provide.
Dependency between Repo and Client in case of using
"dispatching" services allowing the access to entity data
from multiple repositories is NOT removed in this change.

This also makes MultipleRepositoryAwareWikibaseServices implement
EntityStoreWatcher interface, so cache etc. of services from
the container can be updated when changes are done to the entity.

TODO: figure out how to test the "watcher" part of
MultipleRepositoryAwareWikibaseServices

Bug: T165047
Change-Id: Iae1bf00a6b02e8656cf97d27371d9c5ee670ad60