Use WikibaseServices in Client and Repo instead of EntityDataRetrievalServiceFactory
WikibaseServices currently only provide services that
EntityDataRetrievalServiceFactory also used to provide.
Dependency between Repo and Client in case of using
"dispatching" services allowing the access to entity data
from multiple repositories is NOT removed in this change.

This also makes WikibaseServices provide an implementation of
EntityStoreWatcher interface, which can be registered as
a watcher to changes done to the entity.

MultipleRepositoryAwareWikibaseServices returns DispatchingServiceFactory,
that already implements EntityStoreWatcher, as a watcher.

Bug: T165047
Change-Id: Iae1bf00a6b02e8656cf97d27371d9c5ee670ad60