Update patch set 5
Patch Set 5: Code-Review-1

I feel this patch changes to many things the same time, and introduces to many new concepts the same time. Do we really need to do all this right now? For the "statements on forms" use case we have in mind?

I suggest we talk and coordinate first before we continue to work on increasingly complicated patches like this and others I made.

Note that phasing out the wbEntity variable is nothing we can do without close coordination with PM and community manager. There is a huge amount of critical gadgets and user scripts relying on it.

I also suggest to avoid hard coding the idea of "there is only a single entity per page" in more places. As far as I understand this patch it does this and calls this concept "current". What I would love to have instead is a "fetch entity by ID" service that uses the wbEntity variable if it makes sense, and does an API request otherwise.

Patch-set: 5
Label: Code-Review=-1