Update patch set 2
Patch Set 2:

My feeling is that all of ClientStore and much of Store can be folded into WikibaseServices. The remaining code in Store can then perhaps be moved to WikibaseRepo. 

If that is the plan, it would probably be easier to do this before moving them to the new component.

Patch-set: 2