Update patch set 3
Patch Set 3:

I thought about it once more, and I think that this is not good approach. 
If/When we finally remove `wbEntity` from the page, then we will still have statements in HTML, which will take more or less the same amount of space as `wbEntity`.

Another suggestion is to traverse `wbEntity` and find target statement by it's GUID. As soon as we have nested entities we can delegate responsibility of traversing and finding to the Main entity (Lexeme in this case).

Patch-set: 3