Update patch set 5
Patch Set 5:

> I feel this patch changes to many things the same time, and
 > introduces to many new concepts the same time.

The only thing the patch  changes is adds warning message when someone tries to access `wbEntity`. Everything else is new functionality.

 > Do we really need to do all this right now? For the "statements on forms" use case we have in mind?

No. This is almost unrelated. We just talked about it when were discussing
statements on forms and then I had an idea.

 > I also suggest to avoid hard coding the idea of "there is only a
 > single entity per page" in more places. As far as I understand this
 > patch it does this and calls this concept "current". What I would
 > love to have instead is a "fetch entity by ID" service that uses
 > the wbEntity variable if it makes sense, and does an API request
 > otherwise.

It is true that we can have multiple entities on the page, but for now there is always one "main" and 0..N sub-entities. 
If we provide users with ability to request entity by ID then every user-script should request not the last revision (which it will probably do), but the revision that is currently rendered by back-end. Otherwise there might/will be inconsistencies between markup and fetched data and scripts can easily break.
If we provide user-scripts with some wrapper to "current main entity" we can also provide them with method to get sub-entities using the main one. 
May be we should call it differently, not insist on the naming at all.

Patch-set: 5
Label: Code-Review=0