Update patch set 22
Patch Set 22:

How can we help here? When do we want this to be finised? Are we planning to invest non-volunteer resources on this, and if so, when? The ticket T151717 is currently not tracked in the review column of our #Wikidata-Sprint board, and not ready to go there, as far as I can see. The patch is quite nice as it is now, don't get me wrong. The code looks good and we can certainly continue working on this, as Hoo said. But it is still unfinished and not ready to be merged as it is now.

Patch-set: 22
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0