Update patch set 3
Patch Set 3:

@Anomie yea, I was thinking about that. How about I add "my" message to the Status? Would be nice if I could make it the first message instead of the last, though.

The use case for this is: the API code gets a bad Status object that contains error messages, but wants to specify an error code regardless of the specific error. The idea is that the error code specified by the API module would contain more information about the intent of the action, while an error originating somewhere inside the guts of MediaWiki may lack the context necessary for a meaningful interpretation by the user.

Patch-set: 3
Label: Verified=0