Update patch set 2
Patch Set 2:

This patch removes a custom implementation in favor of calling a convenience function, which is fine. However, I think the pull request at https://github.com/wmde/WikibaseDataModelJavaScript/pull/70 does much more than required for this change.

This is how I would suggest to do this instead:

Check if the entity object does have a method called "getStatements". Call it. Check if the returned object is a StatementGroupSet. Use it. That's basically all that's really needed.

For convenience we can add the "getStatementByGuid" function to the StatementGroupSet class.

There is one detail with the approach introducing the "StatementProvider" I did not understood so far: How does code check if an entity implements the StatementProvider interface?

Patch-set: 2