Update patch set 7
Patch Set 7:

I can't seem to find a way to preserve the old-style info message for a known error code. The message is often more specific than the code. 

So, will either have to live with the fact that clients get more generic errors, or I'd have to change the error codes to correspond to more specific messages - which is a breaking change.

@anomie, do you have a nice idea?

Patch-set: 7