Do not fail on item creation in browser tests if the item exists.
Steps like "Given I have an item with label 'foo' and description 'bar'"
fail if the item already exists on the tested wikibase instance which
makes it impossible to run the same test twice. With this change it tries
to create the item and still passes if the item already exists.

Change-Id: Iccecf6ad21220049e76a9bd1e6fe7ea06259d930