Fix MwTimeIsoFormatter hard-coding a bad default date format
This came up in a discussion at T63958. You can see the problem at
https://www.wikidata.org/wiki/Q159?uselang=hu . The date format is
"12 junho 1990", but should be "12. junho 1990" with a dot. This
happens because I hard-coded a date format with the name "dmy", but
there is no guarantee a format with this name exists (even if it does
in the majority of the Message….php files). And the code does not
care about the order. All it cares about is what day and what month
formatting strings are used, and if they end with a dot or comma.

Change-Id: If333e11eb0787189418fb7a36d4ba9a4b61eb28e