Update patch set 6
Patch Set 6:

(1 comment)

Full review done. The actual changes made in this patch look fine. But please note that I'm not able to run ESLint on my Ubuntu 12 installation. I will not be able to contribute any more the same way I was able before.

This is not meant to block anything, just as a note. I will update to Ubuntu 16, but this might take some time. Also I'm curious: what exactly motivated this change? Is there a problem with JSCS and/or JSHint we need to solve?

Patch-set: 6