Update patch set 6
Patch Set 6:

> Also I'm curious: what exactly motivated this change? Is there a problem with JSCS and/or JSHint we need to solve?

Both projects are abandoned, and lack the ability to express some of the nuanced coding standards we have. The former jscs developers have moved to work on eslint, and recommend that we do this (and have been helping us out for the year or so that we've been switching everything over).

Patch-set: 6