Update patch set 28
Patch Set 28:

So as we've just got few of jenkins jobs removed, we have a green light to have Phan checks run as part of regular test build.

In other words: this is ready to go in now!

Apparently just need to wait for VE/Cite tests to be fixed.

Patch-set: 28