Update patch set 7
Patch Set 7: Code-Review+1

I just realized there are 4 different .eslintrc.json in this patch. One in the root, and three in "lib", "repo", and "view", but none in "client". I assume this is meant as a temporary solution. I hope we can make these all the same and remove the additional ones in the subdirectories. Is this impression correct?

Patch-set: 7
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1