Update patch set 73
Patch Set 73:

discussed this with Katie.
Katie will play with the weights to see if the features we have are sufficient to control the ranking in that way. I not we might need to review the features (fields).
Since this process of tuning might take a while it might be interesting to setup a vagrant box running wikibase (using relforge100[12] elastic cluster).

Patch-set: 73
Reviewer: DCausse <2638@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0