Update patch set 7
Patch Set 7:

> I just realized there are 4 different .eslintrc.json in this patch. One in the root, and three in "lib", "repo", and "view", but none in "client". I assume this is meant as a temporary solution. I hope we can make these all the same and remove the additional ones in the subdirectories. Is this impression correct?

Not easily. The normal practice is to apply rule variants at the lowest level possible – so a qunit test suite's access of qunit should be limited to that area and not made more generally available, to avoid qunit being accessed by code that actually runs in production. The split of the test suites into different areas of source code in the way this repo makes that impossible, but the rest of the variation can probably be cleaned up very quickly. Not sure if that usefully answers your question.

Patch-set: 7