Update patch set 2
Patch Set 2: Code-Review+1

Uh. I was afraid we might use the default autosummary when an Item page is turned into a redirect. But I checked and found we are creating our own summary lines in this case. I can not think of something else that might break when this method returns an empty string. But I might miss a lot of things. This needs review by more people.

Patch-set: 2