Add missing check for Language::getDatePreferences return value
This is a direct follow up for a bug introduced in If333e11. The value
returned by Language::getDatePreferences() can be false but is not
documented as such. I trusted the documentation and made this mistake
because of it being incomplete.

I will upload a core patch that fixes the documentation.

Change-Id: Icafe8cb4891c2471c6f30ec4c3c6d27fc89c8ff3