Update patch set 1
Abandoned

Two 1-line private functions turned into one with a parameter. When I created this patch I planned to do more that justifies this better, but did not found enough that was worth refactoring. Also the base patch is still not merged and evolved a lot since I created this. This is just not worth it.

Patch-set: 1
Status: abandoned