Update patch set 5
Patch Set 5: Code-Review-2

According to our policy, a breaking change should be deployed only four weeks after announcement. That means we have to hold back this patch, or add a feature switch, or justify why we don't wait four weeks.

Maybe we can argue that it's breaking, but only "a little bit"?...

Please ask Lydia which option she prefers. Setting this to -2 to prevent merging this without PM decision.

For the record, I think the code is fine, and I agree with the intent. The -2 is just about *when* this can go in, not if.

Patch-set: 5