Update patch set 5
Patch Set 5:

Well, I can make feature switch, but it'd be kind of pointless - why merge it if it's disabled? If we have to wait then we have to wait, there's no point to have this code before we're ready to switch I think.
I'll write to Lydia and ask about her opinion on the break. I think 4 weeks is too long, but I'm ok with waiting for a week - which is, given deployment schedule, two weeks until it's deployed.

Patch-set: 5