Update patch set 1
Patch Set 1: Code-Review-2

This is literally a +/-16,000 lines change. Sixteen thousand. No "git blame" on any code touched in this patch will ever be useful after this is applied. For what benefit? I'm serious: What is the benefit of this change over simply disabling the failing eslint rule? The code style is not like this because somebody made a mistake, but because the people who wrote this code decided to use this code style on purpose. I'm sorry. I believe tools like eslint are exactly that: tools. Tools exist to be useful. Tools are supposed to solve actual problems people have. What is the problem with these 16,000 lines of code, and who does have a problem with it?

Patch-set: 1
Label: Code-Review=-2