Update patch set 1
Patch Set 1:

Perhaps if the indentation was already consistent, but it isn't. In some files the outer closure is indented, in these files it isn't, and in some files is partially indented: either one line or in some cases just half the file. There's no consistency, which I would consider a "mistake". Also you can't disable the "indent" rule for just the outer closure, it has to be completely disabled, so you won't catch indentation errors within the outer closure.

`git blame` isn't an issue, you can use the -w option to ignore whitespace changes, so I see no reason not to fix this.

Patch-set: 1