Update patch set 2
Patch Set 2:

@Bartosz I don't see a corresponding unit test in EntityHandlerTest. 

EditEntityTest tests the behavior of an API module which happens to be influenced by this. That doesn't count as a unit test for the new code.

Sorry to be picky, we have a pretty strict "no new function without a unit test" policy for wikidata. If you don't feel like writing it, maybe I can do it later today. Or maybe I can as Ladsgroup. But I generally don't merge patches without seeing the corresponding unit test work, even if the code is trivial.

Patch-set: 2