Update patch set 2
Patch Set 2:

@Bartosz wrote:
> I will not be adding a test for a method that unconditionally returns the empty string to verify that it always does in fact return the empty string. That's ridiculous.

The purpose of the test is not to check that the current code works correctly. I can see that it does. The idea is to make sure that guaranteed behavior does not change unintentionally later, e.g. due to a mistake during refactoring.

That's the key advantage of high test coverage even for trivial code: confident refactoring. If the tests still pass, I *know* that I didn't break anything while shuffling methods and subclasses and whatever.

The price is writing unit tests for trivial stuff.

Patch-set: 2