Update patch set 1
Patch Set 1:

The code looks fine now, but I'm unsure about whether it's The Right Thing(tm). Auto-summaries are applied when
- a page is turned into a redirect
- when a page is created
- when a page is blanked (or nearly blanked)

For Wikibase, this really doesn't make a difference at the momebnt, since we always provide a summary to our EditEntity class.

But does this mean we need to suppress the AutoSummary? Why? The core change that made this seem necessary has since been improved to not require this, see PS12 on I5041070b47.

Patch-set: 1
Label: Code-Review=0