Update patch set 2
Patch Set 2:

It does not hurt to override this method. If it turns out it will never be called anyway, we can also throw a LogicException instead of returning an empty string, or just remove our override again. All this can easily be done later without the need to block anything.

Patch-set: 2