Change the entire client/includes/… folder to PSR-4
We need to start somewhere and make new code fail if the namespace is
not compatible. The client/includes/… directory is the smallest one
(92 classes) and almost fully converted in previous patches, followed
by lib/includes/… (130), and repo/includes/… (199).

I checked all moved classes if they appear in strings (before …::class
was possible) and @covers tags.

Change-Id: I91800898cebaeb2f27c3f8bf60cab6d5ee6fbf06