Update patch set 5
Patch Set 5:

We talked about this in todays engineering time and decided, as a team, that there is no urge to do this fast. (Please tell us if we are missing something that makes this urgent.) We realize the announcement was already made. We will wait 4 weeks and merge this patch probably a little earlier so it is part of the deployment that will happen 4 weeks after the announcement. Sounds good?

Patch-set: 5