Update patch set 4
Patch Set 4:

Wikidata team has agreed that this change is OK!

So I went ahead and rebased this patch on top of the one introducing eslint config.

@Esanders: one thing that bothers me: how did you actually get all those files tab indented properly? As this is now part of the chain of patches, and some files might have changed since you've uploaded the original patchset, it would probably be the best to simply re-indent all files just before merging this patch, to avoid missing some files, dealing with horrible merge conflicts etc.

When I ran eslint with the fix option, and not "indent" rule not disabled in eslintrc, the tool indeed tab-indented all code as expected but "block comments" were not touched, which results in all those updated files have "unaligned" comments. In turn, in your patch, both code and comments and nicely indented. What am I doing wrong?

Patch-set: 4