Update patch set 4
Patch Set 4: Code-Review+1

(1 comment)

I individually reviewed this patch locally with different diff methods and could not spot an error. This is fine to go in. The only detail missing is an explanation how this patch was created in the first place (see Leszeks comment).

To give some additional context to the -2 I gave originally: years ago the Wikidata team decided to skip the first indention level, because all JavaScript code is wrapped in an anonymous function, which means all lines in a file except the first and last are indented by an extra level. The developers who wanted their code to be like this are not in the team any more. And even if, we are free to change our mind. I just wanted this to be a conscious team decision, which it is now.

Patch-set: 4
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+1