Update patch set 8
Patch Set 8:

(3 comments)

I talked with Aleksey about this quite extensively. I like the overall idea very much.

What this module does is intentionally hard-coding the idea that there is only one entity per page. This concept is called "current entity" in this patch. I'm not sure about the name, but I support the idea.

The alternative would be a more generic module that requires you to specify an entity ID. But where does this ID come from? From the global wbEntityId variable.

The idea presented in this patch allows us to get rid of both wbEntity and wbEntityId. All knowledge about having one entity per page will be encoded in this module, nowhere else. And in case there is no "one entity" on the page, the promise will just not be resolved.

The only alternative idea I would love to play around with is what Daniel said during yesterdays engineering time: copy what MediaWiki core's "wikipage.content" hook does. It is triggered for every content block that finished rendering. The first and only parameter the user gets is the relevant jQuery $content node. The hook is typically triggered when the page loads, one time for the one content block on the page. But it can be triggered any time, multiple times, e.g. every time the user clicks the preview button with the live preview feature enabled. We could do the same. Have a "wikibase.entityview" hook with possibly two parameters: the $entityview node, as well as the entity JSON.

Patch-set: 8
Label: Verified=0