Update patch set 5
Patch Set 5:

I am not really familiar with this part of the code base, so forgive me a possible silly question: where should I look (in terms of tests) to see that the badge is actually added to the side bar?

I see nothing controversial in the code change itself.

Patch-set: 5
Label: Code-Review=0