Update patch set 3
Patch Set 3: Code-Review+1

I have to say this patch is a pain. But since it was created with a tool (I assume, based on the commit message), this is expected, and not a bad thing.

I reviewed this locally using a lot of different greps. The majority of the changes are added spaces to all "function (" headers, as well as added spaces to all "array[ 0 ]" accessors.

I don't want to +2 this myself without a second dev having a look. I would love to get a +2 from either Jonas or Aleksey.

Patch-set: 3