Update patch set 3
Patch Set 3: Code-Review-2

(1 comment)

I was under the impression this was intended to be enabled by default for all users, not only for new users. Is the decision how it should be done documented somewhere? Unfortunately the Phabricator ticket does not link to such a decision.

I'm setting a temporary -2 for the moment. I will also ask Lydia and Lea.

Patch-set: 3