Optimize VariantsAwareRenderer to not output broken weblink syntax
This fixes a user-facing issue when the {{#property:…}} or
{{#statements:…}} parser functions are used to render URLs on a
multi-variant wiki like zhwiki. The generated wikitext is
"-{zh:http://example.com;zh-hans:http://example.com;}-". Try to preview
this wikitext snippet. What happens? The URL parser finds the first
"http://" and eats everything behind, including the semikolons and the
closing "}-".

Instead of trying to fix this combination of the two wikitext features,
I'm skipping the "-{…}-" syntax if all values are the same anyway. I
believe this fixes all issues and is very unlikely to break again in
the future. Let's say the URL renderer changes and starts to output
"[http://example.com <localized label>]" instead. This is fine because
the URL is then guaranteed to end at the "]". The only wikitext able to
break the "-{…}-" sytnax is, as far as I can tell, a plain web link.
But the rendering of a plain URL can not change with the language, and
will always be filtered by the array_unique introduced here.

Bug: T166429
Change-Id: I030f7b375d047f66335a645c549643b59be31bcd